Fixes #2

+++ src/main/java/CalculateService.java
@@ -1,5 +1,5 @@
 public class CalculateService {
-    public int calculate(int fuelAmount, int fuelRate, int distanceUnit) {
-        return fuelAmount * distanceUnit / fuelRate;
+    public int calculate(int fuelAmount, double fuelRate, int distanceUnit) {
+        return (int) (fuelAmount * distanceUnit / fuelRate);
     }
 }
