+++ src/main/java/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         int fuelAmount = 30;
-        int fuelRate = 7;
+        double fuelRate = 7;
         int distanceUnit = 100;
         CalculateService service = new CalculateService();
         System.out.println(service.calculate(fuelAmount, fuelRate, distanceUnit));
